+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Client certificates](https://github.com/makeworld-the-better-one/amfora/wiki/Client-Certificates) can be restricted to certain paths of a host (#115)
 - `header` config option in `[subscriptions]` to allow disabling the header text on the subscriptions page (#191)
 - Selected link and scroll position stays for non-cached pages (#122)
+- Keybinding to open URL with URL handler instead of configured proxy (#143)
 
 ### Changed
 - Center text automatically, removing `left_margin` from the config (#233)